requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
Pillow==11.3.0
ImageHash==4.3.2
piexif==1.1.3
exifread==3.5.1
python-dotenv==1.2.2
requests==2.33.0
opencv-python==4.13.0.92
